+++ soc/CMakeLists.txt
@@ -9,6 +9,8 @@ if(_SOC_IS_IN_TREE)
 endif()
 unset(_SOC_IS_IN_TREE)
 
+add_subdirectory(common)
+
 if(EXISTS ${SOC_DIR}/${ARCH}/CMakeLists.txt)
   add_subdirectory(${SOC_DIR}/${ARCH} soc/${ARCH})
 else()
