+++ soc/common/CMakeLists.txt
@@ -1,0 +1,4 @@
+# Copyright (c) 2024 Nordic Semiconductor ASA
+# SPDX-License-Identifier: Apache-2.0
+
+add_subdirectory_ifdef(CONFIG_SOC_FAMILY_NRF nordic_nrf)
