+++ soc/common/nordic_nrf/CMakeLists.txt
@@ -1,0 +1,4 @@
+# Copyright (c) 2024 Nordic Semiconductor ASA
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_include_directories(.)
