+++ soc/common/nordic_nrf/pinctrl_soc.h
